+++ src/utils/front-matter/index.js
@@ -1,4 +1,4 @@
 export { default as cleanFrontMatter } from "./clean.js";
+export { isEmbedFrontMatter, printEmbedFrontMatter } from "./embed.js";
 export { default as isFrontMatter } from "./is-front-matter.js";
 export { default as parseFrontMatter } from "./parse.js";
-export { default as printEmbedFrontMatter } from "./print.js";
